Extract logic and add test
